refactor(ProductDao): 인터페이스 제거
